refact: removes home page rules and liability
