+++ src/pages/HomePage/viewModel/interface.ts
@@ -1,0 +1,15 @@
+import type { Movie } from '@utils/interfaces';
+
+export interface IHomeViewModel {
+  error: string
+  movies: Movie[]
+  search: string
+  loading: boolean
+  noResult: boolean
+  searchLoading: boolean
+  page: number
+  totalPages: number
+  handleSearchMovies: (text: string) => void
+  loadMoreMovies: () => Promise<void>
+  handleNavigatitonToDetailsPage: (movieId: number) => void
+}
